requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
deep-translator==1.11.4
Django==4.2.10
django-autoslug==1.9.9
django-embed-video==1.4.9
django-modeltranslation==0.18.11
idna==3.6
mysqlclient==2.2.4
pillow==10.2.0
python-dotenv==1.0.1
python-gettext==5.0
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
